Update metadata to change to versions
Jet and I discussed how having all the files in the root of the metadata wasn't a good idea. This changes the format to only allow folders as children of the account. It will also allow folders to be shared (and updated) more easily. It will build off of the account metadata system